Rename scripts and update imports
